refactor: extract create/delete portal element function into utils
